:art: split Utility.swift into several extensions
